Add Nav bar and additional placeholder pages
