+++ src/App.scss
@@ -1,14 +1,88 @@
 @import "@justfixnyc/component-library/dist/src/assets/index.css";
 @import "colors.scss";
 
-body {
-  font-family: "Degular";
-  letter-spacing: 0.54px;
-  background-color: rgba($OFF_WHITE_100, 0.5);
+$headerH: 3.25rem; // 52px
+
+html {
+  height: 100%;
 }
 
-.header-bar {
-  position: absolute;
-  top: 5px;
-  right: 10px;
+body {
+  font-family: "Degular";
+  font-weight: 800;
+  letter-spacing: 0.54px;
+  background-color: rgba($OFF_WHITE_100, 0.5);
+  margin: 0;
+  position: relative;
+  height: 100%;
+}
+
+#header {
+  background-color: $OFF_BLACK;
+  color: $OFF_WHITE_100;
+  font-size: 1.25rem; // 20px
+  letter-spacing: 0.8px;
+  height: $headerH;
+  padding-left: 1rem; // 16px
+  display: flex;
+  align-items: center;
+}
+
+#root {
+  height: 100%;
+}
+
+#main {
+  display: flex;
+  min-height: calc(100% - $headerH);
+}
+
+#sidebar {
+  padding: 24px 16px;
+  border-right: 1px solid #d2d2d2;
+  box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+}
+
+#content {
+  padding: 0 1.5rem; // 24px
+}
+
+#nav {
+  padding: 0 0.75rem; // 0 12px
+}
+
+#nav-links {
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  color: $OFF_BLACK;
+  font-size: 1.25rem; // 20px
+
+  li {
+    margin-bottom: 1rem;
+  }
+  .nav-link {
+    text-decoration: none;
+    color: $OFF_BLACK;
+
+    &:visited {
+      color: $OFF_BLACK;
+    }
+  }
+
+  ul.sublinks {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+
+    li {
+      margin-top: 6px;
+      margin-bottom: 0;
+    }
+
+    .nav-link {
+      font-weight: 300;
+      font-size: 1.125rem; // 18px
+    }
+  }
 }
