+++ src/App.tsx
@@ -1,13 +1,17 @@
 import { Routes, Route, Outlet, Link } from "react-router-dom";
-import { BuildingPage } from "./Components/BuildingPage/BuildingPage";
+import { Buildings } from "./Components/Pages/Buildings/Buildings";
 import { useRollbar } from "@rollbar/react";
 import { SWRConfig } from "swr";
 import { NetworkError } from "./api/error-reporting";
 import { APIDocs } from "./Components/APIDocs/APIDocs";
-import { Login } from "./Components/Login/Login";
-import { PrivateRoutes, useAuth } from "./auth";
+import { Login } from "./Components/Pages/Login/Login";
+import { PrivateRoutes } from "./auth";
 import "./App.scss";
-import { CollectionPage } from "./Components/CollectionPage/CollectionPage";
+import { Landlords } from "./Components/Pages/Landlords/Landlords";
+import { Lenders } from "./Components/Pages/Lenders/Lenders";
+import { Home } from "./Components/Pages/Home/Home";
+import { About } from "./Components/Pages/About/About";
+import { Map } from "./Components/Pages/Map/Map";
 
 function App() {
   const rollbar = useRollbar();
@@ -24,10 +28,13 @@ function App() {
       <Routes>
         <Route path="/" element={<Layout />}>
           <Route element={<PrivateRoutes />}>
-            <Route index element={<>home</>} />
-            <Route path="building" element={<BuildingPage />} />
-            <Route path="collection" element={<CollectionPage />} />
-            <Route path="*" element={<>home</>} />
+            <Route index element={<Home />} />
+            <Route path="buildings" element={<Buildings />} />
+            <Route path="landlords" element={<Landlords />} />
+            <Route path="lenders" element={<Lenders />} />
+            <Route path="about" element={<About />} />
+            <Route path="map" element={<Map />} />
+            <Route path="*" element={<Home />} />
           </Route>
           <Route path="/login" element={<Login />} />
         </Route>
@@ -38,42 +45,74 @@ function App() {
 }
 
 function Layout() {
-  const { user, logout } = useAuth();
   return (
-    <div>
-      <h1>Signature Dashboard</h1>
-      <div className="header-bar">
-        {user && (
-          <>
-            <span className="user-name">Welcome, {user}</span>
-            <button onClick={logout}>Logout</button>
-          </>
-        )}
-        <Link className="api-link" to="/api_docs">
-          API Docs
-        </Link>
-      </div>
-      <nav>
-        <ul>
-          <li>
-            <Link to="/">Home</Link>
-          </li>
-          <li>
-            <Link to="/building?bbl=3071980006">
-              Building Page for 3071980006
-            </Link>
-          </li>
-          <li>
-            <Link to="/collection?collection=ved-parkash">
-              Collection Page for Ved Parkash
-            </Link>
-          </li>
-        </ul>
-      </nav>
+    <>
+      <header id="header">Signature Data Dashboard</header>
+      <div id="main">
+        <div id="sidebar">
+          <nav id="nav">
+            <ul id="nav-links">
+              <li>
+                <Link to="/" className="nav-link">
+                  Home
+                </Link>
+              </li>
+              <li>
+                <Link to="/buildings" className="nav-link">
+                  Buildings
+                </Link>
+              </li>
+              <li>
+                <Link to="/landlords" className="nav-link">
+                  Landlords
+                </Link>
+              </li>
+              <li>
+                <Link to="/lenders" className="nav-link">
+                  Lenders
+                </Link>
+                <ul className="sublinks">
+                  <li>
+                    <Link to="/lenders?lender=cpc" className="nav-link">
+                      CPC
+                    </Link>
+                  </li>
+                  <li>
+                    <Link to="/lenders?lender=santander" className="nav-link">
+                      Santander
+                    </Link>
+                  </li>
+                </ul>
+              </li>
 
-      <hr />
-      <Outlet />
-    </div>
+              <li>
+                <Link to="/about" className="nav-link">
+                  About
+                </Link>
+              </li>
+              <li>
+                <Link to="/map" className="nav-link">
+                  Map
+                </Link>
+              </li>
+              {/* <li>
+                <Link to="/building?bbl=3071980006">
+                  Building Page for 3071980006
+                </Link>
+              </li>
+              <li>
+                <Link to="/collection?collection=ved-parkash">
+                  Collection Page for Ved Parkash
+                </Link>
+              </li> */}
+            </ul>
+          </nav>
+        </div>
+        <div id="content">
+          <Outlet />
+        </div>
+      </div>
+    </>
   );
 }
