+++ src/Components/Collection/CollectionBuildingTable/style.scss
@@ -1,4 +1,4 @@
-@import "./../../../colors.scss";
+@import "../../../colors.scss";
 
 table.collection-building-table,
 table.collection-building-table td,
