+++ src/Components/Collection/CollectionSummaryTable/style.scss
