+++ src/Components/Pages/About/About.tsx
@@ -1,0 +1,11 @@
+// import { AddressRecord } from "../../types/APIDataTypes";
+import React from "react";
+import "./style.scss";
+
+export const About: React.FC = () => {
+  return (
+    <>
+      <h2>About</h2>
+    </>
+  );
+};
