+++ src/Components/Pages/Buildings/style.scss
@@ -1,3 +1,8 @@
+.building-address {
+  font-size: 2.25rem; // 36px
+  margin: 1.5rem 0; // 24px
+}
+
 .chart__timespan_filter {
   display: flex;
