+++ src/Components/Pages/Home/Home.tsx
@@ -1,0 +1,11 @@
+// import { AddressRecord } from "../../types/APIDataTypes";
+import React from "react";
+import "./style.scss";
+
+export const Home: React.FC = () => {
+  return (
+    <>
+      <h2>Home</h2>
+    </>
+  );
+};
