+++ src/Components/Pages/Map/Map.tsx
@@ -1,0 +1,11 @@
+// import { AddressRecord } from "../../types/APIDataTypes";
+import React from "react";
+import "./style.scss";
+
+export const Map: React.FC = () => {
+  return (
+    <>
+      <h2>Map</h2>
+    </>
+  );
+};
